+++ CHANGES.md
@@ -22,6 +22,7 @@ Beta Releases
 * Added `BoundingSphere.fromEllipsoid`.
 * Added renderer support for `OES_element_index_uint`, which can improve performance by reducing batch sizes.
 * Added `czm_tangentToEyeSpaceMatrix` built-in GLSL function.
+* Added `WallGeometry` and `PolygonGeometry`.
 
 ### b17 - 2013-06-03